requirements.txt
pathway[all]>=0.11.0
python-dotenv==1.0.1
llama-index
langchain
langchain_community
fastapi
uvicorn
requests
mpmath==1.3.0
streamlit
openai